==> Competition/2025 동원 x 카이스트 AI Competition Unlocking Future Sales & Demographics/Code/00_0. Data_1.py <==
# import pandas as pd

# # ===== 데이터 입력 (2011년 4분기 ~ 2019년 4분기) =====
# data = {
#     "구분": ["총매출", "백화점", "대형할인마트", "체인슈퍼", "편의점", "독립슈퍼", "일반식품점"],

#     # 2011~2013 (앞에서 정리한 부분)
#     "2011년 4/4분기": [55946, 467, 14112, 14849, 243, 18863, 7412],
#     "2012년 1/4분기": [44329, 461, 11051, 11369, 289, 15013, 6148],
#     "2012년 2/4분기": [42009, 334, 9575, 10500, 274, 14847, 6479],
#     "2012년 3/4분기": [48639, 418, 11967, 12427, 262, 16299, 7265],
#     "2012년 4/4분기": [51697, 398, 11962, 13700, 301, 17532, 7803],
#     "2013년 1/4분기": [39722, 362, 9707, 10312, 246, 13852, 5244],
#     "2013년 2/4분기": [36360, 292, 8117, 9425, 230, 13467, 4829],
#     "2013년 3/4분기": [41272, 310, 9671, 11001, 258, 14819, 5214],
#     "2013년 4/4분기": [46943, 317, 10470, 12607, 295, 17169, 6086],

#     # 2014
#     "2014년 1/4분기": [36663, 337, 7959, 9492, 250, 13676, 4950],
#     "2014년 2/4분기": [36580, 222, 7625, 9524, 262, 14034, 4913],
#     "2014년 3/4분기": [40095, 262, 8628, 10680, 262, 14709, 5554],
#     "2014년 4/4분기": [43900, 288, 8689, 12044, 286, 16450, 6143],

#     # 2015
#     "2015년 1/4분기": [41254, 292, 9478, 10462, 268, 15474, 5279],
#     "2015년 2/4분기": [36343, 238, 7202, 9196, 231, 14498, 4978],
#     "2015년 3/4분기": [39772, 266, 8209, 10304, 259, 15680, 5054],
#     "2015년 4/4분기": [44336, 325, 8432, 12727, 281, 16849, 5722],

#     # 2016
#     "2016년 1/4분기": [41269, 369, 8549, 11271, 287, 15943, 4849],
#     "2016년 2/4분기": [34800, 234, 6392, 9578, 263, 14157, 4175],
#     "2016년 3/4분기": [37520, 279, 7378, 10706, 300, 14697, 4160],
#     "2016년 4/4분기": [42820, 287, 7725, 12320, 348, 17132, 5008],

#     # 2017
#     "2017년 1/4분기": [39755, 315, 8243, 10925, 336, 15642, 4293],
#     "2017년 2/4분기": [35497, 218, 6590, 9766, 326, 14472, 4125],
#     "2017년 3/4분기": [39295, 273, 7555, 11482, 335, 15548, 4102],
#     "2017년 4/4분기": [44209, 268, 8074, 13730, 422, 17413, 4302],

#     # 2018
#     "2018년 1/4분기": [40590, 285, 8326, 11062, 395, 16044, 4478],
#     "2018년 2/4분기": [37112, 243, 6542, 9927, 349, 15596, 4455],
#     "2018년 3/4분기": [39146, 252, 7220, 11169, 368, 16001, 4136],
#     "2018년 4/4분기": [43085, 242, 7398, 12308, 433, 17966, 4736],

#     # 2019
#     "2019년 1/4분기": [40161, 344, 8022, 10666, 405, 16556, 4169],
#     "2019년 2/4분기": [36202, 235, 6680, 9770, 382, 15376, 3760],
#     "2019년 3/4분기": [40337, 239, 7357, 11484, 408, 16859, 3991],
#     "2019년 4/4분기": [43436, 220, 7455, 12794, 436, 18303, 4228],
# }

# df = pd.DataFrame(data)

# # ===== wide → long 변환 =====
# df_long = df.melt(id_vars=["구분"], var_name="기간", value_name="매출액(백만원)")

# # "연도", "분기" 분리
# def parse_period(period):
#     year = int(period.split("년")[0])
#     quarter = int(period.split(" ")[1].split("/")[0])
#     return year, quarter

# df_long[["연도", "분기"]] = df_long["기간"].apply(lambda x: pd.Series(parse_period(x)))

# # 컬럼 순서 정리
# df_final = df_long[["연도", "분기", "구분", "매출액(백만원)"]].sort_values(["연도", "분기", "구분"])

# # ===== 엑셀 저장 =====
# df_final.to_excel("매출_세로형_2011_2019.xlsx", index=False)

# print("엑셀 파일 저장 완료: 매출_세로형_2011_2019.xlsx")


# import pandas as pd

# # ===== 데이터 입력 (2011년 4분기 ~ 2019년 4분기) =====
# data = {
#     "구분": ["총매출", "백화점", "대형할인마트", "체인슈퍼", "편의점", "독립슈퍼", "일반식품점"],

#     # 2011 ~ 2013
#     "2011년 4/4분기": [312655, 2464, 127201, 73086, 16285, 65990, 27629],
#     "2012년 1/4분기": [329433, 2584, 140244, 73934, 18350, 65422, 29439],
#     "2012년 2/4분기": [297609, 2334, 121089, 69864, 13185, 62280, 28857],
#     "2012년 3/4분기": [289563, 2596, 123599, 66325, 10447, 59942, 26653],
#     "2012년 4/4분기": [322300, 2609, 129222, 76225, 17780, 65865, 30599],
#     "2013년 1/4분기": [326718, 2370, 137217, 76929, 17258, 64113, 28830],
#     "2013년 2/4분기": [290685, 2109, 120428, 70490, 12868, 58547, 26243],
#     "2013년 3/4분기": [264997, 1990, 113348, 63748, 9209, 53405, 23296],
#     "2013년 4/4분기": [284088, 1910, 109630, 71080, 15764, 58699, 27004],

#     # 2014
#     "2014년 1/4분기": [292328, 1881, 120757, 71530, 14814, 56927, 26419],
#     "2014년 2/4분기": [259526, 1400, 100375, 67550, 11327, 54839, 24035],
#     "2014년 3/4분기": [244777, 1226, 101703, 60647, 8775, 49302, 23123],
#     "2014년 4/4분기": [259844, 1386, 92875, 69782, 14041, 55270, 26490],

#     # 2015
#     "2015년 1/4분기": [278767, 1267, 115549, 67945, 13478, 55004, 25524],
#     "2015년 2/4분기": [240474, 1155, 88107, 63432, 10806, 53130, 23844],
#     "2015년 3/4분기": [232012, 1047, 94422, 57155, 8007, 50560, 20821],
#     "2015년 4/4분기": [238965, 1182, 82674, 65112, 13122, 53850, 23026],

#     # 2016
#     "2016년 1/4분기": [266000, 1125, 102638, 69253, 13899, 56584, 22502],
#     "2016년 2/4분기": [224838, 927, 82341, 60147, 10897, 50704, 19823],
#     "2016년 3/4분기": [208498, 784, 81261, 55580, 8173, 45248, 17453],
#     "2016년 4/4분기": [230679, 841, 80007, 63482, 14041, 53211, 19097],

#     # 2017
#     "2017년 1/4분기": [250169, 970, 99654, 63967, 13670, 51880, 20028],
#     "2017년 2/4분기": [214949, 781, 78583, 58140, 10656, 49022, 17768],
#     "2017년 3/4분기": [210929, 760, 84603, 56259, 8546, 45212, 15550],
#     "2017년 4/4분기": [225387, 786, 78384, 61031, 14683, 51645, 18859],

#     # 2018
#     "2018년 1/4분기": [242976, 779, 97420, 57737, 14138, 50795, 22107],
#     "2018년 2/4분기": [206685, 658, 72733, 54191, 11281, 47780, 20042],
#     "2018년 3/4분기": [201168, 643, 77996, 52415, 8490, 43950, 17674],
#     "2018년 4/4분기": [222221, 702, 75926, 58985, 13991, 50901, 21716],

#     # 2019
#     "2019년 1/4분기": [225665, 725, 86908, 53368, 13532, 49383, 21749],
#     "2019년 2/4분기": [193500, 574, 66436, 50366, 10839, 46299, 18987],
#     "2019년 3/4분기": [193619, 552, 76488, 48444, 8482, 43203, 16450],
#     "2019년 4/4분기": [195227, 563, 62124, 52554, 13040, 47155, 19792],
# }

# df = pd.DataFrame(data)

# # ===== wide → long 변환 =====
# df_long = df.melt(id_vars=["구분"], var_name="기간", value_name="매출액(백만원)")

# # "연도", "분기" 분리
# def parse_period(period):
#     year = int(period.split("년")[0])
#     quarter = int(period.split(" ")[1].split("/")[0])
#     return year, quarter

# df_long[["연도", "분기"]] = df_long["기간"].apply(lambda x: pd.Series(parse_period(x)))

# # 컬럼 순서 정리
# df_final = df_long[["연도", "분기", "구분", "매출액(백만원)"]].sort_values(["연도", "분기", "구분"])

# # ===== 엑셀 저장 (지정 경로) =====
# save_path = r"C:\python\Study25_집\_data\dacon\dongwon\pos_data\커피_2011_2019_버전2.xlsx"
# df_final.to_excel(save_path, index=False)

# print(f"엑셀 파일 저장 완료: {save_path}")

# import pandas as pd

# # ===== 데이터 입력 (2020년 상반기 ~ 2024년 하반기) =====
# data = {
#     "구분": ["총매출", "백화점", "대형할인마트", "체인슈퍼", "편의점", "독립슈퍼", "일반식품점"],

#     # 2020
#     "2020년 상반기": [404331, 875, 109291, 103505, 25249, 102620, 62791],
#     "2020년 하반기": [383603, 844, 110932, 97726, 23253, 93087, 57760],

#     # 2021
#     "2021년 상반기": [386330, 768, 112967, 97022, 25352, 91022, 59200],
#     "2021년 하반기": [343117, 0, 93693, 87197, 22333, 85033, 54860],

#     # 2022
#     "2022년 상반기": [357536, 0, 95954, 88366, 24317, 89627, 59272],
#     "2022년 하반기": [353633, 0, 100234, 90033, 22715, 85681, 54971],

#     # 2023
#     "2023년 상반기": [368275, 0, 98224, 94517, 27628, 89155, 58751],
#     "2023년 하반기": [351889, 0, 99355, 92694, 24702, 83137, 52001],

#     # 2024
#     "2024년 상반기": [350262, 0, 101016, 88956, 29906, 82585, 47799],
#     "2024년 하반기": [333015, 0, 95918, 87772, 27736, 78530, 43058],
# }

# df = pd.DataFrame(data)

# # ===== wide → long 변환 =====
# df_long = df.melt(id_vars=["구분"], var_name="기간", value_name="매출액(백만원)")

# # "연도", "반기" 분리
# def parse_period(period):
#     year = int(period.split("년")[0])
#     half = 1 if "상반기" in period else 2
#     return year, half

# df_long[["연도", "반기"]] = df_long["기간"].apply(lambda x: pd.Series(parse_period(x)))

# # 컬럼 순서 정리
# df_final = df_long[["연도", "반기", "구분", "매출액(백만원)"]].sort_values(["연도", "반기", "구분"])

# # ===== 엑셀 저장 (지정 경로) =====
# save_path = r"C:\python\Study25_집\_data\dacon\dongwon\pos_data\닐슨코리아 커피_2020_2024.xlsx"
# df_final.to_excel(save_path, index=False)

# print(f"엑셀 파일 저장 완료: {save_path}")


# import pandas as pd

# # ===== 데이터 입력 (2020년 1월 ~ 2023년 12월) =====
# data = {
#     "구분": ["총매출", "백화점", "대형할인마트", "체인슈퍼", "편의점", "독립슈퍼", "일반식품점"],

#     # 2020
#     "2020년01월": [83454, 342, 26180, 18837, 6651, 22719, 8725],
#     "2020년02월": [75600, 346, 22200, 17208, 6391, 20976, 8479],
#     "2020년03월": [97979, 408, 28817, 22552, 8173, 26934, 11095],
#     "2020년04월": [80793, 317, 24352, 18098, 6467, 22593, 8966],
#     "2020년05월": [83760, 269, 22704, 20024, 5561, 25563, 9639],
#     "2020년06월": [92394, 327, 27495, 22240, 5805, 26612, 9915],
#     "2020년07월": [68244, 264, 20992, 16096, 4264, 19485, 7142],
#     "2020년08월": [66564, 242, 21443, 15714, 3758, 18359, 7048],
#     "2020년09월": [89142, 366, 27672, 20799, 5747, 24789, 9769],
#     "2020년10월": [73136, 275, 20910, 17133, 5291, 21359, 8166],
#     "2020년11월": [77330, 281, 23481, 17864, 5576, 21710, 8418],
#     "2020년12월": [117883, 447, 34326, 27129, 9013, 33572, 13397],

#     # 2021
#     "2021년01월": [72604, 299, 21095, 15579, 6123, 20959, 8550],
#     "2021년02월": [72990, 252, 20499, 16232, 5846, 21589, 8572],
#     "2021년03월": [88432, 347, 27282, 18947, 7255, 24846, 9754],
#     "2021년04월": [66671, 250, 20134, 14783, 4970, 18992, 7543],
#     "2021년05월": [64969, 250, 19744, 14471, 4867, 18309, 7327],
#     "2021년06월": [78272, 339, 25221, 18089, 5258, 20986, 8379],
#     "2021년07월": [55343, 233, 17949, 12376, 3641, 14994, 6150],
#     "2021년08월": [53232, 188, 17062, 11037, 3801, 15048, 6096],
#     "2021년09월": [79925, 272, 23082, 17853, 5489, 23914, 9316],
#     "2021년10월": [62182, 211, 18118, 14041, 5299, 17585, 6927],
#     "2021년11월": [67600, 226, 19540, 15969, 5492, 18963, 7411],
#     "2021년12월": [85258, 298, 24823, 18645, 7087, 24654, 9750],

#     # 2022
#     "2022년01월": [72420, 268, 22070, 16011, 5613, 20247, 8210],
#     "2022년02월": [66377, 242, 18866, 13903, 5968, 19301, 8096],
#     "2022년03월": [86276, 301, 27789, 17479, 6506, 24210, 9990],
#     "2022년04월": [68323, 234, 20598, 14521, 5213, 19866, 7891],
#     "2022년05월": [66782, 246, 20681, 14562, 4748, 18928, 7617],
#     "2022년06월": [79504, 294, 25336, 17710, 5395, 22058, 8711],
#     "2022년07월": [60096, 206, 19304, 13068, 3990, 16852, 6675],
#     "2022년08월": [58608, 199, 18481, 12653, 4044, 16431, 6800],
#     "2022년09월": [81277, 280, 24094, 17220, 6226, 23922, 9535],
#     "2022년10월": [69099, 220, 20277, 14312, 5630, 20343, 8318],
#     "2022년11월": [74177, 230, 22135, 16266, 5793, 21011, 8742],
#     "2022년12월": [94799, 301, 28519, 19174, 7773, 27224, 11809],

#     # 2023
#     "2023년01월": [74730, 239, 21666, 15745, 6114, 21528, 9439],
#     "2023년02월": [74059, 252, 21326, 15369, 6212, 21497, 9403],
#     "2023년03월": [93316, 288, 29567, 19753, 7486, 25165, 11056],
#     "2023년04월": [70952, 240, 21649, 15791, 5809, 19315, 8148],
#     "2023년05월": [67246, 240, 20332, 15201, 5266, 18182, 8025],
#     "2023년06월": [80291, 261, 25314, 18764, 5863, 20962, 9127],
#     "2023년07월": [60939, 223, 18790, 13732, 4519, 16542, 7133],
#     "2023년08월": [57916, 176, 18554, 13481, 4026, 15305, 6374],
#     "2023년09월": [82816, 234, 24043, 18973, 6006, 23842, 9719],
#     "2023년10월": [67239, 214, 19735, 15091, 5646, 18644, 7907],
#     "2023년11월": [73339, 216, 22525, 16592, 6266, 19523, 8216],
#     "2023년12월": [90041, 268, 26024, 20235, 7925, 25088, 10501],
# }

# df = pd.DataFrame(data)

# # ===== wide → long 변환 =====
# df_long = df.melt(id_vars=["구분"], var_name="기간", value_name="매출액(백만원)")

# # "연도", "월" 분리
# def parse_period(period):
#     year = int(period[:4])
#     month = int(period[5:7])
#     return year, month

# df_long[["연도", "월"]] = df_long["기간"].apply(lambda x: pd.Series(parse_period(x)))

# # 컬럼 순서 정리
# df_final = df_long[["연도", "월", "구분", "매출액(백만원)"]].sort_values(["연도", "월", "구분"])

# # ===== 엑셀 저장 (지정 경로) =====
# save_path = r"C:\python\Study25_집\_data\dacon\dongwon\pos_data\마켓링크 커피_2020_2023_월별.xlsx"
# df_final.to_excel(save_path, index=False)

# print(f"엑셀 파일 저장 완료: {save_path}")


# import pandas as pd

# # ==========================
# # 어육가공품 데이터 입력
# # ==========================
# data = {
#     "구분": ["총매출","백화점","대형할인마트","체인슈퍼","편의점","독립슈퍼","일반식품점"],
#     "2020-01":[55350,705,15341,13882,7843,13442,4137],
#     "2020-02":[50482,701,13368,12255,7986,12317,3854],
#     "2020-03":[72467,975,18763,18885,10006,18208,5630],
#     "2020-04":[51722,722,12980,12928,8057,12917,4117],
#     "2020-05":[49339,637,11485,11413,8838,12756,4209],
#     "2020-06":[57371,782,13272,12931,11156,14520,4711],
#     "2020-07":[45640,640,11174,10581,8587,11096,3561],
#     "2020-08":[49137,731,12918,11454,8473,11753,3809],
#     "2020-09":[68408,968,17552,16856,11069,16719,5243],
#     "2020-10":[50974,721,13372,12324,8239,12432,3884],
#     "2020-11":[48721,712,12830,11935,7481,11992,3770],
#     "2020-12":[84824,1183,23101,21559,11703,20770,6509],
#     "2021-01":[57769,829,15419,13927,8846,14360,4389],
#     "2021-02":[51228,673,13352,12252,8502,12668,3780],
#     "2021-03":[57932,822,15318,13557,10661,13463,4112],
#     "2021-04":[44116,645,11473,10350,8131,10276,3241],
#     "2021-05":[44964,630,11205,10259,9216,10316,3338],
#     "2021-06":[54164,784,14029,12079,11384,11996,3892],
#     "2021-07":[43387,598,11609,9882,9081,9218,3000],
#     "2021-08":[47544,674,12611,11090,9676,10210,3284],
#     "2021-09":[62095,821,15161,14708,12598,14234,4572],
#     "2021-10":[47714,656,11808,11562,9765,10573,3350],
#     "2021-11":[46834,651,12184,11562,8342,10681,3412],
#     "2021-12":[63215,843,17474,15458,10822,14136,4481],
#     "2022-01":[56772,751,15397,13913,9886,12981,3844],
#     "2022-02":[47095,615,12087,10521,9709,10839,3324],
#     "2022-03":[61391,865,16802,14141,11288,14077,4219],
#     "2022-04":[45351,625,11602,9739,9816,10359,3210],
#     "2022-05":[44747,612,11467,9576,10140,9785,3167],
#     "2022-06":[56481,787,14860,11961,12749,12225,3899],
#     "2022-07":[45406,619,12196,9481,10099,9877,3134],
#     "2022-08":[50046,649,13953,10611,10703,10782,3347],
#     "2022-09":[67029,789,18427,14854,14045,14432,4483],
#     "2022-10":[52455,683,14077,11732,10349,11972,3642],
#     "2022-11":[51289,651,13926,11730,9361,11900,3721],
#     "2022-12":[69202,911,20455,15897,11143,15918,4878],
#     "2023-01":[57055,686,16766,14021,8522,13102,3958],
#     "2023-02":[48708,637,13277,11457,8186,11481,3670],
#     "2023-03":[60794,784,17860,14012,10923,12972,4243],
#     "2023-04":[46879,626,12709,10924,9029,10296,3294],
#     "2023-05":[45999,614,12515,10471,9255,9979,3166],
#     "2023-06":[55712,772,15252,12667,11781,11547,3693],
#     "2023-07":[44460,625,12527,9825,9430,9171,2881],
#     "2023-08":[46405,640,13573,10530,9502,9199,2962],
#     "2023-09":[65453,810,18187,15773,12129,14079,4476],
#     "2023-10":[47253,622,13186,11264,9238,9778,3165],
#     "2023-11":[49591,664,14505,12184,8442,10408,3388],
#     "2023-12":[64928,864,20345,16061,10071,13263,4324],
# }

# # ==========================
# # DataFrame 생성
# # ==========================
# df = pd.DataFrame(data)

# # ==========================
# # 엑셀 저장
# # ==========================
# output_path = "어육가공품_매출_2020_2023.xlsx"
# df.to_excel(output_path, index=False)

# print(f"엑셀 파일이 저장되었습니다 → {output_path}")


# import pandas as pd
# from pathlib import Path

# # ===== 경로 =====
# DATA_DIR = Path(r"C:\python\Study25_집\_data\dacon\dongwon\pos_data")
# SAVE_PATH = DATA_DIR / "어육가공품_매출_2011_2019.xlsx"

# # ===== 데이터 입력 =====
# data = {
#     "구분": ["총매출", "백화점", "대형할인마트", "체인슈퍼", "편의점", "독립슈퍼", "일반식품점"],
#     "2011년4분기": [99176, 1445, 32396, 24285, 6108, 29255, 5687],
#     "2012년1분기": [104606, 1474, 33764, 25753, 6759, 30351, 6505],
#     "2012년2분기": [95198, 1311, 29127, 22335, 7114, 28501, 6810],
#     "2012년3분기": [107095, 1319, 32843, 25049, 8644, 32085, 7155],
#     "2012년4분기": [115882, 1459, 35183, 28048, 8630, 34681, 7881],
#     "2013년1분기": [114809, 1462, 34791, 28757, 7910, 34312, 7577],
#     "2013년2분기": [101379, 1362, 29893, 24983, 7915, 30520, 6705],
#     "2013년3분기": [105205, 1387, 31829, 25528, 9141, 30708, 6613],
#     "2013년4분기": [114160, 1544, 34107, 28288, 8570, 34154, 7497],
#     "2014년1분기": [107141, 1432, 32426, 26925, 7313, 32132, 6912],
#     "2014년2분기": [94585, 1156, 27514, 23567, 7355, 28600, 6394],
#     "2014년3분기": [99287, 1262, 30437, 24396, 8139, 28874, 6180],
#     "2014년4분기": [110714, 1554, 34226, 27248, 8269, 32556, 6860],
#     "2015년1분기": [104837, 1329, 32616, 26450, 7369, 30607, 6467],
#     "2015년2분기": [96412, 1211, 28821, 23609, 8018, 28426, 6328],
#     "2015년3분기": [102804, 1192, 31586, 24874, 8827, 29883, 6443],
#     "2015년4분기": [113777, 1371, 34707, 29737, 9282, 31846, 6835],
#     "2016년1분기": [112728, 1296, 33311, 30414, 8836, 32529, 6343],
#     "2016년2분기": [97608, 1145, 28879, 25086, 8911, 28106, 5482],
#     "2016년3분기": [104542, 1127, 31719, 26521, 11086, 28624, 5465],
#     "2016년4분기": [118356, 1279, 35097, 29958, 12972, 32837, 6213],
#     "2017년1분기": [116155, 1222, 34806, 30051, 11441, 32663, 5973],
#     "2017년2분기": [100523, 1004, 29620, 25517, 11248, 27905, 5228],
#     "2017년3분기": [111141, 1081, 34251, 28137, 13423, 29205, 5043],
#     "2017년4분기": [118729, 1170, 36590, 30640, 13128, 31756, 5445],
#     "2018년1분기": [120560, 1083, 36917, 30734, 12793, 32101, 6932],
#     "2018년2분기": [109169, 975, 32102, 26774, 14538, 28504, 6278],
#     "2018년3분기": [112782, 997, 34253, 27948, 15363, 28048, 6173],
#     "2018년4분기": [118566, 1081, 35497, 30620, 14466, 30232, 6671],
#     "2019년1분기": [110761, 970, 33690, 29037, 12409, 28711, 5945],
#     "2019년2분기": [99883, 896, 30031, 25870, 12626, 25145, 5315],
#     "2019년3분기": [106571, 906, 32888, 28046, 13314, 26304, 5113],
#     "2019년4분기": [115918, 971, 37196, 30826, 13572, 27611, 5743],
# }

# # ===== DataFrame 변환 =====
# df = pd.DataFrame(data)

# # ===== 저장 =====
# df.to_excel(SAVE_PATH, index=False)

# print(f"저장 완료: {SAVE_PATH}")


# import pandas as pd
# from pathlib import Path

# # ===== 경로 =====
# DATA_DIR = Path(r"C:\python\Study25_집\_data\dacon\dongwon\pos_data")
# SAVE_PATH = DATA_DIR / "어육가공품_매출_2020_2024.xlsx"

# # ===== 데이터 입력 =====
# data = {
#     "구분": ["총매출", "백화점", "대형할인마트", "체인슈퍼", "편의점", "독립슈퍼", "일반식품점"],
#     "2020년 상반기": [257150, 2118, 63365, 69364, 28869, 71690, 21944],
#     "2020년 하반기": [267781, 2392, 72599, 69366, 32277, 70084, 21062],
#     "2021년 상반기": [255476, 2172, 70985, 66162, 29143, 66945, 20069],
#     "2021년 하반기": [242950,    0, 64582, 63512, 33498, 62987, 18371],
#     "2022년 상반기": [237115,    0, 64147, 63723, 29040, 62529, 17676],
#     "2022년 하반기": [256747,    0, 74131, 69442, 30748, 64995, 17431],
#     "2023년 상반기": [248573,    0, 71334, 70215, 27793, 62942, 16288],
#     "2023년 하반기": [263451,    0, 80025, 74868, 28167, 64678, 15714],
#     "2024년 상반기": [247466,    0, 75959, 71227, 27432, 58939, 13909],
#     "2024년 하반기": [260727,    0, 78512, 76501, 29902, 61601, 14212],
# }

# # ===== DataFrame 변환 =====
# df = pd.DataFrame(data)

# # ===== 저장 =====
# df.to_excel(SAVE_PATH, index=False)

# print(f"저장 완료: {SAVE_PATH}")
