==> Competition/2025 동원 x 카이스트 AI Competition Unlocking Future Sales & Demographics/Code/00_0. Data_7_Price list.py <==
import pandas as pd

# 제품별 시세 데이터 구성
data = [
    ["덴마크 하이그릭요거트 400g", "400g", 3980, "9.95원/g"],
    ["동원맛참 고소참기름 135g", "135g", 2380, "17.6원/g"],
    ["동원맛참 고소참기름 90g", "90g", 2380, "26.4원/g"],
    ["동원맛참 매콤참기름 135g", "135g", 3980, "29.5원/g"],
    ["동원맛참 매콤참기름 90g", "90g", 2380, "26.4원/g"],
    ["동원참치액 순 500g", "500g", 5700, "11.4원/g"],
    ["동원참치액 순 900g", "900g", 9980, "11.1원/g"],
    ["동원참치액 진 500g", "500g", 5700, "11.4원/g"],
    ["동원참치액 진 900g", "900g", 9980, "11.1원/g"],
    ["프리미엄 동원참치액 500g", "500g", 10700, "21.4원/g"],
    ["프리미엄 동원참치액 900g", "900g", 22980, "25.5원/g"],
    ["리챔 오믈레햄 200g", "200g", 5080, "25.4원/g"],
    ["리챔 오믈레햄 340g", "340g", 7480, "22.0원/g"],
    ["소화가 잘되는 우유로 만든 바닐라라떼 250mL", "250mL", 2680, "10.7원/mL"],
    ["소화가 잘되는 우유로 만든 카페라떼 250mL", "250mL", 2680, "10.7원/mL"],
]

# DataFrame 생성
df_prices = pd.DataFrame(data, columns=["제품명", "용량", "대표 판매가", "단위당 가격(참고)"])

# 엑셀 파일 저장
df_prices.to_excel("신제품_시세_가격표.xlsx", index=False, engine="openpyxl")

print("엑셀 파일 저장 완료: 신제품_시세_가격표.xlsx")
